% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.survnet}
\alias{predict.survnet}
\title{survnet prediction}
\usage{
\method{predict}{survnet}(object, newdata, cause = NULL, ...)
}
\arguments{
\item{object}{\code{survnet} object}

\item{newdata}{New data predictors: \code{matrix}, \code{array} or \code{data.frame}.}

\item{cause}{Select cause for competing risks, \code{NULL} returns list of all causes.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Cumulative incidence function of selected or all causes.
}
\description{
survnet prediction
}
