% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survnet.R
\name{survnet}
\alias{survnet}
\title{Artificial neural networks for survival analysis}
\usage{
survnet(
  y,
  x,
  breaks,
  units = c(3, 5),
  units_rnn = c(4, 6),
  units_causes = c(3, 2),
  epochs = 100,
  batch_size = 16,
  validation_split = 0.2,
  loss = loss_cif_loglik,
  activation = "tanh",
  rnn_type = "LSTM",
  skip = TRUE,
  dropout = rep(0, length(units)),
  dropout_rnn = rep(0, length(units_rnn)),
  dropout_causes = rep(0, length(units_causes)),
  l2 = rep(0, length(units)),
  l2_rnn = rep(0, length(units_rnn)),
  l2_causes = rep(0, length(units_causes)),
  optimizer = optimizer_rmsprop(learning_rate = 0.001),
  verbose = 2
)
}
\arguments{
\item{y}{Survival outcome: \code{matrix}, \code{data.frame} or \code{Surv()} object.}

\item{x}{Predictors: \code{matrix}, \code{data.frame} or \code{array} (time-series). Also accepts a list of \code{matrix/data.frame} and \code{array} for both time-constant and time-series predictors.}

\item{breaks}{Right interval limits for discrete survival time.}

\item{units}{Vector of units, each specifying the number of units in one hidden layer.}

\item{units_rnn}{Vector of units for recurrent layers.}

\item{units_causes}{Vector of units for cause-specific layers (competing risks only). Either a vector (will be repeated for each cause) or a list of vectors with layers for each cause.}

\item{epochs}{Number of epochs to train the model.}

\item{batch_size}{Number of samples per gradient update.}

\item{validation_split}{Fraction in [0,1] of the training data to be used as validation data.}

\item{loss}{Loss function.}

\item{activation}{Activation function.}

\item{rnn_type}{Type of RNN layers. Either \code{"LSTM"} (default) or \code{"GRU"}.}

\item{skip}{Add skip connection from input and RNN layers to cause-specific layers.}

\item{dropout}{Vector of dropout rates after each hidden layer. Use 0 for no dropout (default).}

\item{dropout_rnn}{Vector of dropout rates after each recurrent layer. Use 0 for no dropout (default).}

\item{dropout_causes}{Vector of dropout rates after each cause-specific layer. Use 0 for no dropout (default).}

\item{l2}{Vector of L2 regularization factors for each hidden layer. Use 0 for no regularization (default).}

\item{l2_rnn}{Vector of L2 regularization factors for each recurrent layer. Use 0 for no regularization (default).}

\item{l2_causes}{Vector of L2 regularization factors for each cause-specific layer. Use 0 for no regularization (default).}

\item{optimizer}{Name of optimizer or optimizer instance.}

\item{verbose}{Verbosity mode (0 = silent, 1 = progress bar, 2 = one line per epoch).}
}
\value{
Fitted model.
}
\description{
Artificial neural networks for survival analysis
}
\examples{
\donttest{
library(survival)
library(survnet)

# Survival data
y <- veteran[, c(3, 4)]
x <- veteran[, c(-2, -3, -4)]
x <- data.frame(lapply(x, scale))
breaks <- c(1, 50, 100, 200, 500, 1000)

# Fit simple model
fit <- survnet(y = y, x = x, breaks = breaks)
plot(fit$history)
}

}
