% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{loss_cif_loglik}
\alias{loss_cif_loglik}
\title{Cumulative incidence log-likelihood}
\usage{
loss_cif_loglik(num_intervals, num_causes = 1)
}
\arguments{
\item{num_intervals}{Number of time intervals}

\item{num_causes}{Number of causes for competing risks}
}
\value{
Negative log-likelihood
}
\description{
Likelihood of parametric inference for the cumulative incidence functions as defined by Jeong & Fine 2006. Also used by Lee et al. 2018.
}
\details{
Data structure:

\code{y_true} True survival: Matrix with at-risk and event information. Format: (S_1, ..., S_K, E_1, ..., E_K). Dimensions: obs X 2*causes*time.

\code{y_pred} Network output: One probability for each time and cause. Format: (y_11, ..., y_T1, ..., y_TK). Dimensions: obs X causes*time.
}
\references{
\itemize{
  \item Jeong, J. & Fine, J. (2006). Direct parametric inference for the cumulative incidence function. J R Stat Soc Ser C Appl Stat 55:187-200. \url{https://doi.org/10.1111/j.1467-9876.2006.00532.x}.
  \item Lee, C., Zame, W.R., Yoon, J. & van der Shaar, M. (2018). DeepHit: A deep learning approach to survival analysis with competing risks. AAAI 2018. \url{http://medianetlab.ee.ucla.edu/papers/AAAI_2018_DeepHit}.
 }
}
