% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survnet.R
\name{survnet}
\alias{survnet}
\title{Artificial neural networks for survival analysis}
\usage{
survnet(y, x, breaks, units = c(3, 5), units_rnn = c(4, 6),
  units_causes = c(3, 2), epochs = 100, batch_size = 16,
  validation_split = 0.2, loss, activation = "tanh",
  optimizer = optimizer_rmsprop(lr = 0.001), plot = TRUE)
}
\arguments{
\item{y}{Survival outcome, created with \code{Surv()}.}

\item{x}{Predictors: \code{matrix}, \code{array} or \code{data.frame}.}

\item{breaks}{Right interval limits for discrete survival time.}

\item{units}{Vector of units, each specifying the number of units in one hidden layer.}

\item{units_rnn}{Vector of units for recurrent layers.}

\item{units_causes}{Vector of units for cause-specific layers (competing risks only). Either a vector (will be repeated for each cause) or a list of vectors with layers for each cause.}

\item{epochs}{Number of epochs to train the model.}

\item{batch_size}{Number of samples per gradient update.}

\item{validation_split}{Fraction in [0,1] of the training data to be used as validation data.}

\item{loss}{Loss function.}

\item{activation}{Activtion function.}

\item{optimizer}{Name of optimizer or optimizer instance.}

\item{plot}{Plot training and validation errors.}
}
\value{
Fitted model.
}
\description{
Artificial neural networks for survival analysis
}
